+++ lib/pixar_ruby_extensions/string/predicates.rb
@@ -14,6 +14,9 @@ module PixarRubyExtensions
       INTEGER_RE = /\A-?[0-9]+\Z/.freeze
       FLOAT_RE = /\A-?[0-9]+\.[0-9]+\Z/.freeze
 
+      DELIM_PLACEHOLDER = '_DELIM_'
+      UUID_RE_STR = "\\A[a-f0-9]{8}#{DELIM_PLACEHOLDER}([a-f0-9]{4}#{DELIM_PLACEHOLDER}){3}[a-f0-9]{12}\\z"
+
       # Does this string contain an integer?
       # (i.e. it consists only of numeric digits,
       #  maybe with a dash in front)
@@ -43,6 +46,23 @@ module PixarRubyExtensions
         true if Float(self)
       rescue
         false
+      end
+
+      # Is this string a UUID?
+      #
+      # Looks for the standard hyphenated form: 32 hexadecimal characters (0-9, a-f)
+      # and four hyphen delimiters in this pattern: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
+      #
+      # You can specify a different delimiter if needed.
+      #
+      # @param delimiter [String] the delimiter used in the UUID, default '-'
+      #
+      # @return [Boolean]
+      #
+      def pix_uuid?(delimiter = '-')
+        re_str = UUID_RE_STR.gsub DELIM_PLACEHOLDER, Regexp.escape(delimiter)
+        regex = Regexp.new re_str, 'i'
+        regex.match? self
       end
 
     end # module